refine utils
