+++ simrec/utils/distributed.py
@@ -95,3 +95,13 @@ def reduce_meters(meters, rank, cfg):
             if main_process(cfg, rank):
                 value = th.mean(th.cat(avg_reduce)).item()
                 meter.update_reduce(value)
+
+def find_free_port():
+    import socket
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    # Binding to port 0 will cause the OS to find an available port for us
+    sock.bind(("", 0))
+    port = sock.getsockname()[1]
+    sock.close()
+    # NOTE: there is still a chance the port could be taken by other processes.
+    return port
